Move all existing round trip Go tests to `integration`
The test code has been moved as verbatim as possible. There will be some cleanups in follow-up PRs. The only changes made in this PR are those that were necessary for linting to pass.
